chore(design-system): move ComponentPage to pages directory
Moves ComponentPage out of the ui-components folder and into the pages folder. Also updates the imports in the routes to point to the new location.
