+++ plugins/design-system/routes/ui-components/text-fields.js
@@ -1,6 +1,6 @@
 import { navigation, routing } from "foundation-ui";
 
-import ComponentPage from "../../pages/ui-components/ComponentPage";
+import ComponentPage from "../../pages/ComponentPage";
 
 import OverviewTab
   from "../../pages/ui-components/text-fields/tabs/OverviewTab";
